requirements.txt
python-telegram-bot[job-queue]==21.6
cryptography==42.0.8
APScheduler==3.10.4
psycopg[binary]==3.2.1
psycopg-pool==3.2.3
